Split SimpleFlatTableProducer plugin into header and implementation
